+++ web/client/components/season2/vote/step3.vue
@@ -126,7 +126,7 @@
 </template>
 
 <script>
-import voteErrorMessages from '~/data/season2/vote-error-messages'
+import voteErrorMessages from '~/data/common/vote-error-messages'
 
 export default {
   name: 'VoteStep3',
