+++ web/client/data/common/vote-error-messages.js
